Feito verificação se o devedor tem vínculo com devedores antes de excluir

+++ src/devedores/devedores.service.ts
@@ -1,12 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { ClienteDevedor } from 'src/clientes-devedores/cliente-devedor.entity';
 import { Repository } from 'typeorm';
 import { Devedor } from './devedor.entity';
 
 @Injectable()
 export class DevedoresService {
     constructor(
-        @InjectRepository(Devedor) private devedoresRepository: Repository<Devedor>
+        @InjectRepository(Devedor) private devedoresRepository: Repository<Devedor>,
+        @InjectRepository(ClienteDevedor) private clienteDevedorRepository: Repository<ClienteDevedor>
     ) {}
 
     async findAll(): Promise<Devedor[]> {
@@ -27,15 +29,45 @@ export class DevedoresService {
         return await this.devedoresRepository.findOne(id);
     }
 
-    async remove(id: string): Promise<boolean> {
-        const response = await this.devedoresRepository.delete(id);
+    async remove(id: string): Promise<any> {
+        const hasDebit = await this.devedorHasDebit(Number(id));
 
-        if(response) {
-            return true;
+        if(!hasDebit) {
+            const response = await this.devedoresRepository.delete(id);
+
+            if(response) {
+                return {
+                    status: 'success',
+                    message: 'Devedor excluído com sucesso.'
+                }
+            }
+            else {
+                return {
+                    status: 'error',
+                    message: 'Problemas ao excluir devedor.'
+                }
+            }
         }
         else {
-            return false;
+            return {
+                status: 'error',
+                message: 'Não foi possível excluir, pois este devedor tem vínculo com débitos de devedores.'
+            }
         }
     }
 
+    async devedorHasDebit(devedorId: number): Promise<boolean> {
+        const hasDebit = await this.clienteDevedorRepository.find({
+            where: {
+                devedor: devedorId
+            }
+        });
+
+        if(hasDebit.length == 0) {
+            return false;
+        }
+        else {
+            return true;
+        }
+    }
 }
